backend-services/functions/trigger-monitor/requirements.txt
functions-framework==3.*
cloudevents==1.10.*
google-cloud-aiplatform>=1.60.0
google-cloud-firestore>=2.14.0
